factor out region and evaluation_type into their own files
