+++ nrcan_etl/src/energuide/embedded/evaluation_type.py
@@ -1,0 +1,15 @@
+import enum
+
+@enum.unique
+class EvaluationType(enum.Enum):
+    PRE_RETROFIT = 'D'
+    POST_RETROFIT = 'E'
+
+    @classmethod
+    def from_code(cls, code: str) -> 'EvaluationType':
+        if code == cls.PRE_RETROFIT.value:
+            return EvaluationType.PRE_RETROFIT
+        elif code == cls.POST_RETROFIT.value:
+            return EvaluationType.POST_RETROFIT
+        else:
+            raise InvalidInputDataError(f'Invalid code: {code}')
