+++ nrcan_etl/src/energuide/embedded/region.py
@@ -1,0 +1,41 @@
+import enum
+import typing
+
+@enum.unique
+class Region(enum.Enum):
+    BRITISH_COLUMBIA = 'BC'
+    ALBERTA = 'AB'
+    SASKATCHEWAN = 'SK'
+    MANITOBA = 'MB'
+    ONTARIO = 'ON'
+    QUEBEC = 'QC'
+    NEW_BRUNSWICK = 'NB'
+    PRINCE_EDWARD_ISLAND = 'PE'
+    NOVA_SCOTIA = 'NS'
+    NEWFOUNDLAND_AND_LABRADOR = 'NL'
+    YUKON = 'YT'
+    NORTHWEST_TERRITORIES = 'NT'
+    NUNAVUT = 'NU'
+    UNKNOWN = '??'
+
+    @classmethod
+    def _from_name(cls, name: str) -> typing.Optional['Region']:
+        snake_name = name.upper().replace(' ', '_')
+        return Region[snake_name] if snake_name in Region.__members__ else None
+
+    @classmethod
+    def _from_code(cls, code: str) -> typing.Optional['Region']:
+        code = code.upper()
+        for region in Region:
+            if code == region.value:
+                return region
+        return None
+
+    @classmethod
+    def from_data(cls, data: str) -> 'Region':
+        output = cls._from_name(data)
+        if not output:
+            output = cls._from_code(data)
+        if not output:
+            output = Region.UNKNOWN
+        return output
