+++ nrcan_etl/tests/test_database.py
@@ -1,6 +1,7 @@
 import typing
 import pymongo
 import pytest
+from energuide.embedded import region
 from energuide import database
 from energuide import dwelling
 
@@ -13,7 +14,7 @@ def load_data() -> typing.List[dwelling.Dwelling]:
             house_type='Single detached',
             year_built=2000,
             city='Ottawa',
-            region=dwelling.Region.ONTARIO,
+            region=region.Region.ONTARIO,
             forward_sortation_area='K1P',
             evaluations=[]
         ),
@@ -22,7 +23,7 @@ def load_data() -> typing.List[dwelling.Dwelling]:
             house_type='Single detached',
             year_built=2000,
             city='Ottawa',
-            region=dwelling.Region.ONTARIO,
+            region=region.Region.ONTARIO,
             forward_sortation_area='K1P',
             evaluations=[]
         ),
@@ -31,7 +32,7 @@ def load_data() -> typing.List[dwelling.Dwelling]:
             house_type='Single detached',
             year_built=2000,
             city='Ottawa',
-            region=dwelling.Region.ONTARIO,
+            region=region.Region.ONTARIO,
             forward_sortation_area='K1P',
             evaluations=[]
         ),
@@ -71,7 +72,7 @@ def test_load_update(database_coordinates: database.DatabaseCoordinates,
         house_id=1,
         house_type='Single detached',
         year_built=2001, city='Ottawa',
-        region=dwelling.Region.ONTARIO,
+        region=region.Region.ONTARIO,
         forward_sortation_area='K1P',
         evaluations=[]
     )
@@ -81,7 +82,7 @@ def test_load_update(database_coordinates: database.DatabaseCoordinates,
             house_type='Single detached',
             year_built=2001,
             city='Ottawa',
-            region=dwelling.Region.ONTARIO,
+            region=region.Region.ONTARIO,
             forward_sortation_area='K1P',
             evaluations=[]
         )
